Updated openstack/openstack Project: openstack/cinder  2752afe461c7e45532d447f85c940fa9771e4360
api-ref: Add backup import and export

During the migration to in-tree api-ref docs we somehow lost the details
for backup export_record and import_record. Adding back in.

Also fixing up a few records that referred to the wrong parameter record
names, resulting in their descriptions referring to volume transfer for
backup operations.

Change-Id: I6fc555729c1b8404887d41424831fc48934f1491
Closes-bug: #1628812
Closes-bug: #1712923