Updated openstack/openstack Project: openstack/tempest  344bdffd620f730bc946c8f6752233ecdd90625d
Test shared and non-shared external networks

The external network visibility test verifies the subnet visibility of
an external network.  The current implementation only works correctly
for a non-shared, external network.  In that case the test verifies that
the subnets are not visible.

For a shared, external network the test incorrectly fails.  This change
verifies the subnet visibility as follows:
 * if the external network is shared, the subnets are visible
 * if the external network is not shared, the subnets are not visible

Change-Id: I5bacf2f0081f659c84fe3920687e2da61bab6237
Closes-Bug: #1704467