Updated openstack/openstack Project: openstack/nova  be3a66781f7fd58e5c5c0fe89b33f8098cfb0f0d
Clarify the field usage guidelines

The note explaining not to use the Nova versions of object field types
mentions that they will be removed, implying that the types listed
there, such as Dict and List should not be used. This update clarifies
that it is only the Nova versions of these fields that will be removed,
and that it is perfectly fine to use the o.vo field types.

Change-Id: I4ae2392e97f5cbe59e5c9c94c06c48a5d9c9d7d3