Updated openstack/openstack Project: openstack/tripleo-heat-templates  43692d1fccdf34c50300b016fc6219aca7630ec2
Use switch to containers instead of take over playbook for ceph-ansible

On upgrade we need to run a specific playbook for ceph-ansible
to be able to take over the pre-existing Ceph cluster deployed with
puppet-ceph and the migrate it into a containerized deployment.

This changes the playbook we use on upgrade so that it migrates
the cluster in containers in addition to taking over the cluster.

Change-Id: I353c219832c41328f298fa7b65768ecf26c37f29