Updated openstack/openstack Project: openstack/tripleo-heat-templates  ecc3f69778815f2a583fecfc16580d4ad79464bf
Change ManageEventPipeline to true

Since the event pipeline publisher defaults in the heat templates are
different from what puppet sets. We need to have the Manage to true so
override takes effect. Without this we keep defaulting back to puppet
defaults. We can flip this back to false once panko:// is droppped as
a supported option from the pipeline.

Change-Id: I2248c165783dddfb4cb7cf5644884dd8f6e6ed63