Updated openstack/openstack Project: openstack/tripleo-heat-templates  1f2e349dcad1fdf5f265af5b76c929094b7d2bc9
Add clustercheck to service list for scenarios

This service is necessary when we containerized TripleO with
Pacemaker.

The service is added also to non-containerized scenario lists, because
the aim is to get rid of the -containers.yaml variants eventually.
This shouldn't affect any jobs that don't include docker-ha.yaml. The
resource registry entry is mapped to OS::Heat::None by default, and
docker-ha.yaml maps it to actual containerized clustercheck.

Change-Id: I342e29de52cb6ce069a05a2dbfb0501a2da200e6
Partial-Bug: #1712070