Updated openstack/openstack Project: openstack-infra/project-config  2bc665caf2dcefa690074f7136dbdf82b75a3d71
Add comment to top of zuul.yaml pointing to other files

Zuul v3 config is spread across a bunch of repos, so figuring out where
somethig is defined may be daunting for a newcommer. Add a comment to
the top of the file pointing people at other repos they should be aware
of.

Change-Id: Ie6ce5e277d93cf348e707c69a3979d8543ba36c6