Updated openstack/openstack Project: openstack-infra/project-config  bd1ba5b4c56048b96cf67a6f6334570fcd6ba05c
Add bindep_command and bindep_fallback to site-variables

The bindep role is set up to allow setting a bindep command and fallback
file via variables, and to do the right thing if they're not there.

Add bindep_command and bindep_fallback to our site-variables so that we
can remove copy-pasta references from jobs.

Change-Id: I94dfe8e6aab9c2da1283590be5e26f7d8c35dcad