Updated openstack/openstack Project: openstack/cinder  7eccc7dbc208982f0d498bcfa540e0aba1d3511d
[Doc] Handle Sphinx warnings as errors

We should be treating warnings from Sphinx as
errors as they usually result in bad HTML output.

Now that we have all of the existing warnings resolved we
want to turn on this option so that more bad docstrings, etc.
do not get merged.

One last bad docstring appears to have snuck in.  This patch
also fixes that docstring problem in cinder/volume/qos_specs.py .

We have a requirement on graphviz that was never handled and is
actually causing one of our web pages to be improperly rendered.
The patch adds the dependency to bindep.txt so that we do not
see a check failure and so the documentation builds properly.

Change-Id: I66a9282b63d884a390534b0c777be8433a7e9bb3