Updated openstack/openstack Project: openstack-infra/project-config  b6bb5bfc045cf187cab95b0fd03eab24fa99bb82
Add a non-voting and reliable tripleo job to devstack-gate check queue

We've landed a change in devstack-gate that broke the tripleo gate
yesterday. Since devstack-gate will be undergoing extensive changes
as we move towards ansibilization and zuul v3, let's add a non-voting
job to try and prevent this from re-occurring.

Change-Id: Ifc667e3af097739ff74b77665785c2b6ac85afce
Related-Bug: #1713868