Updated openstack/openstack Project: openstack-infra/shade  cd572f273350c537b1d148fec696c69b5d9e2dc5
Fix handling timeouts in volume functional tests cleanup

This patch adds handle for fixture.TimeoutException in same way
as it was done for OpenStackCloudTimeout.

It also changes cleanup function that it will not wait for
delete of volumes if volume was stuck in "deletion" state.
Delete volume with "force=True" will now be call with
"wait=False"

Change-Id: Ie6a8c7aecaf36b07247b61296d68afd923280440