Updated openstack/openstack Project: openstack-infra/shade  69e88470832dde44b2d6d440e97182257158a875
Add handling timeout in servers cleanup function

In functional tests in method which cleans server and
volumes attached to it, there was sometimes reached timeout,
probably because one of resources stuck in "deletion" state
for longer than 180 seconds.
This patch adds handling such timeout and in such case,
try to call delete of server and each volume but without
waiting for deletion to be completed and with "force" option
in case of volumes.

Change-Id: I7a8caf2e89348ae0931a7e7e30ce2a83c4d12e5a