Updated openstack/openstack Project: openstack-infra/devstack-gate  e69d8b835db3783301fd33e04acadf1d06d307e4
Revert "Revert "Move fix_disk_layout to an Ansible role in setup_host""

This re-introduces the Ansible role to set up the disk layout.
The initial introduction increased dramatically the occurrence of the
race condition involving the failure when configuring swap on /dev/xvde1 [1].

This (re)revert substitutes the sync and partprobe commands by a
udevadm settle command in order to wait until the device and it's partitions
are ready for use.

[1]: https://bugs.launchpad.net/openstack-gate/+bug/1706103

This reverts commit 71aa1bd6cfe7e7a98179008638e27f2f4c675441.

Change-Id: Iac0efdb6077ab18c2e1991820da4fcea7642f708