Updated openstack/openstack Project: openstack/tempest  f1fbb3fd8a29c1a61ebd32a14134511ec27cf4da
Fix identity tests when domain specific drivers are enabled

This updates users and groups identity tests to check the existence
of domain specific drivers and, therefore, the existence of users
and groups on different domain drivers.

This adds a new feature flag to be used when domain specific drivers
are enabled.

Change-Id: Iedb470c51fa2174ab7651e6b7e22eff1f25f7aac