Updated openstack/openstack Project: openstack/tempest  5ee9e19c8828b88e8c6c1fd7f491f16be8aa3074
Fix the test case for showing host

From the implementation of Cinder show host API, we can know that the API
is to show volume usage info on the specified cinder-volume host, which is
not fully described in official api-ref. If the specified host does not run
the cinder-volume service, or the cinder-volume service is disabled on the
host, the API should fail. This patch aims to solve this problem.

Change-Id: I40e64c52a5a71d3a7fd7e7f04b86b6c35d49594e
Closes-Bug: #1691144