Updated openstack/openstack Project: openstack/nova  a4609a834b1eaf5013f7da8580628ecd1d51227d
api-ref: add warnings about forcing the host for live migrate/evacuate

Bypassing the scheduler during a live migration or evacuation
is not something we want people to be doing as it can easily fail
if the specified host is already full or doesn't provide something
required by the instance, plus it's a nightmare for tracking allocations
in the Placement service when we're bypassing the scheduler.

Because of this, we should have some warnings in the API reference
about doing this, which this patch adds.

Change-Id: I85e7c2677f4d5eccc1e7f349de06960b53ef148d