Updated openstack/openstack Project: openstack/tripleo-heat-templates  9dfa1a2e0f1b1529d97ca571ad48dbd855ec9047
Add storage backends env files for containerized deployment

A storage backend has to be selected when deploying manila,
otherwise the manila-share service will fail to start. For this,
we have some environment files specifying the configuration for
different storage backends. We need a dockerized version
for this environment files.

In this patch set we add those environment files.

Change-Id: I9886016b02bec26699af1f8165d7b0702dfe8b9b
Partial-Bug: #1668922