Updated openstack/openstack Project: openstack/nova  beac1529fdb92dbf17994f785d058fd5b413c588
doc: link to versioned notification samples from main index

Notifications are essentially another API for end users, but
it can be hard to find the list of existing notifications or
their samples which are buried deep down via contributors >
technical reference deep dive > notitfications. If I'm an
end user, I'd like to just see them in the same set of links
on the main page as the API, under the "For End Users" section.

Change-Id: If3ca21b080d06a291ed27c9bcd84a566164c3b70