Updated openstack/openstack Project: openstack/tripleo-heat-templates  7ce354336554eb2362b60c082b4df6ea4b760cad
Remove tacker from containers scenario001

The containerized implementation of tacker is incomplete in THT,
and relies on the pre-pike single "tacker" container. Container
builds using the final pike release of kolla build three tacker
containers to have seperate conductor and server containers.

According to this bug[1], tacker does not even work without this
conductor. Our scenario job needs to be updated to actually test
tacker is working.

This will need to be backported to pike, and we can work on
better supporting tacker in containers in queens.

[1] https://bugs.launchpad.net/tripleo/+bug/1710874

Change-Id: I7cab33687a05bf6ba5c6fb70ba21f3250d3ef381
Partial-Bug: 1714270