Updated openstack/openstack Project: openstack/tripleo-heat-templates  4b789f4963649c0356ed4ef10197e83a9a81796a
Allow upgrade tasks to run when looping through steps

Currently for non controller upgrades we're looping through the
upgrade steps and run the upgrade tasks based on when conditionals
including the step number and the existing upgrade task condition.
Some of tasks fail because the variables used in when conditionals
are not available through all steps. This change adds default values
to these vars where possible or creates them for all steps to avoid
failures.

Related-Bug: 1708115
Change-Id: I5c731043cec8e31fc82ca98972a301baa7294c4f