Updated openstack/openstack Project: openstack/nova  be223198f6a41e288acfa3a34cc06a4879988184
Move hash ring initialization to init_host() for ironic

The hash ring initialization code requires calling to conductor over
RPC. Nova's service startup initializes the manager and the virt
driver, then goes into a polling loop waiting for a conductor to
respond to make sure it is ready. Finally, it calls init_host()
once things are expected to be working. That is where we should
do our initialization that requires conductor to be up.

This patch moves the initial refresh into the Ironic driver's
init_host() routine.

Change-Id: If92f4b1df061f91ffffa7900532a978649efb933
Closes-Bug: #1713095