Updated openstack/openstack Project: openstack/nova  e795c8a09bcb447695034b0b4a1a8f8f23a0edb2
Fix scope of errors_out_migration in resize_instance

We currently execute the whole of resize_instance in the context of
the errors_out_migration decorator, meaning that any error will result
in the migration objection being set to error. However, we cast the
migration process to the destination compute before sending local
notification before sending instance notifications from the source
host. This means that an error sending instance notifications on the
source compute would set the error state of a migration which is
currently executing on a different compute host. Setting error here
isn't appropriate, as well as being a race.

Here we reduce the scope of errors_out_migration to end immediately
after the cast to the destination compute.

In adding a test for the reduced scope, we also refactor
test_resize_instance_failure to reuse its suite of resize_insetance
mocks. We take the opportunity to remove unnecessary mocks and
simplify others, and to remove assertions which don't relate to code
correctness or what we're testing.

Change-Id: Id2f352a68f99df430d20634695be51f74db95d87