Updated openstack/openstack Project: openstack/cinder  0ab70a9f4b00e493c67018b9a324ca99541bc1a2
Make sure that versions are applied as relative URLs

When urljoining a base_url with a version, the path in the base_url will
be dropped if the version string starts with a "/". This will lead to
issues when the base_url looks like "https://HOSTIP/volume/" like in the
devstack tls-proxy setup.

Change-Id: Ia2060066330d8e7dd40f37392db4d6b7b9a6fa8e
Partial-Bug: 1713732