Updated openstack/openstack Project: openstack/governance  e698cf66033d2727b5dea8636857010d6bc81dc3
Drop Technical Committee meetings

So, this is an idea I've been poking around for a bit. TBH, I ran this
idea through some folks last year but it didn't seem to be the right
moment for it. During the last elections, Matt asked[0] a question about
inclusiveness in the TC, which resulted in myself[1] and ttx[2]
proposing this idea.

[0] http://lists.openstack.org/pipermail/openstack-dev/2017-April/115227.html
[1] http://lists.openstack.org/pipermail/openstack-dev/2017-April/115255.html
[2] http://lists.openstack.org/pipermail/openstack-dev/2017-April/115313.html

Now, down to business. From a practical perspective, most of the
discussions we have at a TC level happen (or should happen) on gerrit
and the mailing list. Lately, most of our meetings have been mostly
useful to rubber stamp things that we've voted on already on gerrit.

This is not to say that meetings are useless and that we don't have any
interesting discussions during these meetings. We do, in fact, have some
more heated discussions during our meetings from time to time but they
are not as freequent as they used to be.

So, instead of having a blocked hour, every week, for the TC meeting,
I'd like to propose we drop the meetings entirely and, instead,
encourage discussion in #openstack-dev. Here's are some things I think
this would give us:

- Possibility for arranging more focused discussions with stackholders
  on hours that work best. The smaller the group, the easier it is to
  pick a time.

- More open and transparent discussions as it'll force everyone to
  discuss things using gerrit and the mailing list instead of waiting
  for meetings to happen.

- It'll be easier for folks to stay up-to-date with the on-going
  discussions as we'll be reducing one channel of communication and
  focus on the other 2 we're already using.

- It'll be easier for folks (especially non-native English speakers) to
  participate in the discussions as it'll allow them to take their time
  to do it.

Co-Authored: John Garbutt <john.garbutt@rackspace.com>
Change-Id: If70bb95da3216aad7632983f1955d06de1c50521