Updated openstack/openstack Project: openstack/api-site  551cfeeeec1bda2a6884231b3eb8af9b30d7820a
Redirect shared-file-systems to shared-file-system

The official service type for Manila is shared-file-system (singular).
The official API reference should be standardized accordingly, and in
fact, this is the case ([1] is following the latest commits; [2] stopped
being updated on 20170718).

This change accordingly updates static links within this project, and
adds 301 redirects for anything pointing to the old path.

[1] https://developer.openstack.org/api-ref/shared-file-system/
[2] https://developer.openstack.org/api-ref/shared-file-systems/

Change-Id: I24019199bed3eb4eb7030bf49fb04a910a674675