Updated openstack/openstack Project: openstack-infra/devstack-gate  b91005fc521d4881998c8147b8dd89ad42e04e65
Allow for arguments to DEVSTACK_GATE_SETTINGS.

The following patch added some arguments to be passed to the
DEVSTACK_GATE_SETTINGS file:

    https://review.openstack.org/394684

This is not currently supported by devstack-gate, as the argument is
included as part of the filename when checking to see if the settings
file exists.  This patch makes it ignore arguments when checking for
the existence of the settings file.

Closes-bug: 1665856
Change-Id: Ie787baeeafd38670e4085d734329b64eff3d3741
Signed-off-by: Russell Bryant <rbryant@redhat.com>