Updated openstack/openstack Project: openstack/nova  dbfde14978d4e4374c1fc1085c7f061b2a22d2c6
Add recreate test for forced host evacuate not setting dest allocations

This adds a functional recreate test for bug 1713786 where
allocations are not created against the forced destination host
during an evacuation.

This asserts the faulty behavior so that a later patch which fixes
the bug can uncomment the assertions for correct behavior.

Change-Id: I55bf64a95f6b1b781656920380dc38bda864bd14
Related-Bug: #1713786