Updated openstack/openstack Project: openstack/nova  15eebf4dc18d8f6d117a1a66c64ff1abc4a37932
Cleanup allocations on invalid dest node during live migration

Starting in Pike, the scheduler creates an allocation on a
chosen destination node during live migration. This happens
before the destination node pre-checks occur, which could
fail and trigger a retry to the scheduler. The allocations
created in Placement against the failed destination node
were not being cleaned up though.

This change adds some cleanup code to the live migration task
in conductor to clean the allocations for the failed destination
node before retrying.

The functional recreate test for the bug is updated to show
that the bug is fixed now.

Also updates the docstring in the SchedulerReportClient
remove_provider_from_instance_allocation method so that we
no longer have to enumerate all of the places that call it.

Change-Id: I41e5e1fa9938b5e04f7e20f78ccd77eca658885f
Closes-Bug: #1712411