Updated openstack/openstack Project: openstack-dev/devstack  b9e368b9f60778bef0c6ddea122ca020a1f760b1
Add procname for uwsgi based services

Code in grenade and elsewhere rely on the process/service name
when one runs "ps auxw" and they grep for example "grep -e glance-api"
to check if the service is running. with uwsgi, let us make sure
we use process name prefix so it is easier to spot the services
and be compatible with code elsewhere that relies on this.

Change-Id: I4d1cd223ed9904fcb19b26fc9362b676e0b4f9b3