Updated openstack/openstack Project: openstack/neutron  765f76ae3651405ccf2638a1d0e51cddc7b8de13
dvr: Don't raise KeyError in _get_floatingips_bound_to_host

We thought _get_floatingips_bound_to_host is not needed but removing the
method caused sending garps for fip that doesn't belong to node during
the full-sync.

This patch just replaces dict lookup with get() method, so fips are
filtered based on presence on the host and if host is not set on fip, it
won't raise a KeyError.

Co-Authored-By: Swaminathan Vasudevan <SVasudevan@suse.com>

Related-bug: #1712412
Related-bug: #1713927

Change-Id: I0fbc772d757fb13b788f9df8d6d7d28d288d054a