Updated openstack/openstack Project: openstack-infra/project-config  041391e84967ae2d4251c3f7fe4a5b13a6b5a15b
Update puppet-ceph to use stable/pike for jewel

We have recently cut a release for the jewel version of the module which
is supported until Pike. The current master is going to be updated for
the next LTS of ceph so we need to test the pike modules on the jewel
branch.

Change-Id: I8fbe3b56b17d1db6f1d4cbaa37e338943470d56c