Updated openstack/openstack Project: openstack/neutron  c11a4038d984b84a00e4d9030ec2c8b6e85a9aae
Pecan: set tenant_id field when project_id set

Add logic to pecan to add 'project_id' as a mandatory policy
field whenever 'tenant_id' is specified as a policy field.
This logic was added to the previous controller code to ensure
that 'project_id' was queried in the database whenever 'tenant_id'
was required by the policy engine and the user had fields set.

Closes-Bug: #1714378
Change-Id: I3652fbc50ce0c9a7cd1cc193e0933cf0373ecb54