Updated openstack/openstack Project: openstack/diskimage-builder  3ff8d1e10b5ade89b0081bb39f8ab6dd921dc0c8
Move the ordering of the dracut regenerate command

When using combined with rhel7 image, the unregister of repos
has already happened, because it is executed under 60- ordering.
As dracut-regenerate may need to install extra packages for it,
it causes this step to fail, because it cannot find repos where
to pull the packages from.

Change-Id: I35e37df7990ad76a5004cb90fdd863ec743a5483