Updated openstack/openstack Project: openstack/nova  0a36453a81b827bb4d3f4ee39626109129f6c821
Glance download: only fsync files

Recent changes [1][2] added fsync to the data file in
GlanceImageServiceV2.download.  This raises EINVAL if the file is a
pipe/FIFO or socket [3].

This change set adds a static _safe_fsync method to GlanceImageServiceV2
which conditions the fsync call not to run if the file handle represents
a pipe/FIFO or socket, and uses that call from the download method.

[1] https://review.openstack.org/#/c/441246/
[2] https://review.openstack.org/#/c/443583/
[3] http://man7.org/linux/man-pages/man2/fsync.2.html#ERRORS

Change-Id: Ied5788deadcf3d1336a48288cf49d8571db23659
Closes-Bug: #1677047