Updated openstack/openstack Project: openstack/nova  823b118b8cbdd3e7a0db87da0031d4e5f3f2c864
trivial: Remove files from 'tools'

Two files are removed.

'ebtables.workaround' was required for libvirt < 1.2.11 for the reasons
described in the file itself. We no longer support this version of
libvirt, thus, the workaround can be removed.

'regression_tester' is a tool for running tests in a patch against old
code, to ensure the tests validate what the patch is supposed to be
fixing. This is unmaintained and is not referenced anywhere, and should
therefore be removed as it's likely bitrotted (or will do so).

Change-Id: Iac7f16ff4c178b5e71b786c332ca46a78bef5e3c
TrivialFix