Updated openstack/openstack Project: openstack/swift  7e5a39a5bdb6334dca935dc6d5e6d4bf790a2800
Add explanation about Content-Length header limit with DLO to docs

If a number of DLO segments is larger than container listing limit,
Content-Length header will not be included in GET or HEAD response.
However, this fact is not explained in document of large objects.

This patch add explanation about this fact to the document.

Change-Id: Ia45fad05797f38fa8b6b0ed917b4f9d7fb337149
Closes-Bug: 1680219